# DOT-LMS-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOT-LMS API Documentation",
        "contact": {
            "email": "ruthgetaneh5@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "DOT-LMS API Server"
        }
    ],
    "paths": {
...
        "/Events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of Events in the DOT-LMS",
                "description": "Returns list of Events",
                "operationId": "9519bf96ff99c219930a3ad56ea69736",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Store new Event's data into the database",
                "description": "Returns the entered Event's data",
                "operationId": "178ce997a94463a6a92a004d9f39d47e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Events"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/Events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get a specific Event",
                "description": "Returns the specified Event's data",
                "operationId": "a5f50c2d7ab3ee2f27d9067c9a34de1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EVNT-6832"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update existing Events",
                "description": "Returns updated Event's data",
                "operationId": "7bd07086408c8bcb8597a8cab7625c0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EVNT-6832"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Events"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete existing Event",
                "description": "Deletes a record and returns no content",
                "operationId": "ca27b0c4f61501f6c8f6094193ffd980",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Event's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "EVNT-683"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/Events/Teacher/{teacher_id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Event based on a teacher's ID",
                "description": "Get Event data based on a teacher's ID",
                "operationId": "c399824cda26c4c6e12396bbf22bb65d",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "description": "Teacher's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "TCH-6823"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/Events/Course/{course_id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get Event based on a course's ID",
                "description": "Get Event data based on a course's ID",
                "operationId": "938902390ec28c51bade931c72f1c755",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "Course's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "CUS-6823"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Events": {
                "title": "Events",
                "properties": {
                    "event_title": {
                        "title": "event_title",
                        "description": "The title of the Event",
                        "type": "string",
                        "example": "Assignment Deadline"
                    },
                    "event_start": {
                        "title": "event_start",
                        "description": "The starting date of the event",
                        "type": "string",
                        "format": "date"
                    },
                    "event_end": {
                        "title": "event_end",
                        "description": "The ending date of the event",
                        "type": "string",
                        "format": "date"
                    },
                    "teacher_id": {
                        "title": "teacher_id",
                        "description": "The id of the teacher that created the event",
                        "type": "string",
                        "example": ""
                    },
                    "course_id": {
                        "title": "course_id",
                        "description": "The id of the course the event belongs to",
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "api_key",
                "in": "header"
            }
        }
    },
...
}